pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inboxdna"
dynamic = ["version"]
description = "Local-first Gmail inbox analyzer and cleanup tool"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Blaise Ratcliffe" },
]
keywords = ["gmail", "email", "inbox", "cleanup", "privacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
]
dependencies = [
    "Flask>=3.0",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.0",
    "google-auth-httplib2>=0.2",
]

[project.urls]
Homepage = "https://github.com/blaiseratcliffe/inboxdna"
Issues = "https://github.com/blaiseratcliffe/inboxdna/issues"

[project.scripts]
inboxdna = "inboxdna.app:main"

[tool.setuptools.packages.find]
include = ["inboxdna*"]

[tool.setuptools.dynamic]
version = {attr = "inboxdna.__version__"}

[tool.setuptools.package-data]
inboxdna = [
    "templates/*.html",
    "static/**/*",
    "client_secret_*.json",
]
